refactor: Improve deleteItem function in firebase.ts and ListItem.tsx. The deleteItem function in firebase.ts has been updated to handle the case where the itemDocRef is null, improving the robustness of the delete functionality. Additionally, the deleteItem function in ListItem.tsx now uses a try-catch block for error handling and displays an alert message on error. If an error occurs during the deletion process, it will be caught and logged to the console. Furthermore, the deleteItem function in firebase.ts has been modified to display the deleted item name in the alert message. This change enhances the overall delete functionality and user experience. Co-authored-by: Maha Ahmed <eternalmaha@users.noreply.github.com>

+++ src/api/firebase.ts
@@ -297,16 +297,20 @@ export async function updateItem(listPath: string, item: ListItem) {
 
 //delete an item from the list
 export async function deleteItem(listPath: string, item: ListItem) {
-	const itemDocRef = doc(db, listPath, "items", item.id) || null;
+	const itemDocRef = doc(db, listPath, "items", item.id);
+	const docSnap = await getDoc(itemDocRef);
+
 	// Delete the item from the list in Firestore.
 	// Should be a simple try-catch block. Try === deletes the item. Catch === logs the error.
 	// Let's try the deleteDoc from Firestore.
-	try {
-		await deleteDoc(itemDocRef);
-		alert(`${item.name} has been successfully deleted!`);
-	} catch (error) {
-		// If there's an error, log it to the console and throw it.
-		console.error("Oops! Error deleting Item!", error);
-		throw error;
+	if (docSnap.exists()) {
+		try {
+			await deleteDoc(itemDocRef);
+			alert(`${item.name} has been successfully deleted!`);
+		} catch (error) {
+			// If there's an error, log it to the console and throw it.
+			console.error("Oops! Error deleting Item!", error);
+			throw error;
+		}
 	}
 }
